corrected id type

+++ ArticleManagement/ArticleManagement/src/main/java/nwtprojekat/ArticleManagement/repository/SectionRepository.java
@@ -3,7 +3,7 @@ package nwtprojekat.ArticleManagement.repository;
 import nwtprojekat.ArticleManagement.model.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectionRepository extends JpaRepository<Section,Integer> {
+public interface SectionRepository extends JpaRepository<Section, Long> {
 
 
 }
